requirements.txt
# 核心评测与API请求依赖
mteb>=1.1.0
pymilvus>=2.3.0
python-dotenv>=1.0.0
numpy>=1.24.0
tqdm>=4.65.0
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.14.0
jieba>=0.42.1

# 可选：根据你的embedding模型选择
# sentence-transformers>=2.2.0
# 或
# openai>=1.0.0
# anthropic>=0.18.0
